refactor: 추상 클래스를 통한 service 중복 제거
