+++ tomcat/src/main/java/nextstep/jwp/servlets/RootServlet.java
@@ -4,21 +4,12 @@ import static org.apache.coyote.http11.message.HttpHeaders.CONTENT_LENGTH;
 import static org.apache.coyote.http11.message.HttpHeaders.CONTENT_TYPE;
 
 import org.apache.coyote.http11.ContentType;
-import org.apache.coyote.http11.message.HttpMethod;
 import org.apache.coyote.http11.message.HttpStatus;
 import org.apache.coyote.http11.message.request.HttpRequest;
 import org.apache.coyote.http11.message.response.HttpResponse;
 import org.apache.coyote.http11.message.response.ResponseBody;
 
-public class RootServlet extends Servlet {
-
-    @Override
-    public void service(HttpRequest httpRequest, HttpResponse httpResponse) {
-        HttpMethod httpMethod = httpRequest.getMethod();
-        if (httpMethod.isEqualTo(HttpMethod.GET)) {
-            doGet(httpRequest, httpResponse);
-        }
-    }
+public class RootServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpRequest httpRequest, HttpResponse httpResponse) {
